+++ internal/signals_unix.go
@@ -1,0 +1,12 @@
+//go:build aix || darwin || dragonfly || freebsd || linux || netbsd || openbsd || solaris
+// +build aix darwin dragonfly freebsd linux netbsd openbsd solaris
+
+package internal
+
+import (
+	"syscall"
+)
+
+func init() {
+	TerminationSignals = append(TerminationSignals, syscall.SIGTERM)
+}